✨ Add method to check for tag support
The datadog statsd client supports tags but the regular one does not. We have cases where we want to add tags in shared libraries that causes errors in services that do not use the datadog client.

Adding a method to report on the tag support of the instantiated client
so that we can guard these cases.

See SLAPI-278